Refactor: postDetail 헤더의 게시글 삭제/수정/신고를 postItem으로 옮겨 구현 (#87)
